requirements.txt
neo4j==5.15.0
sentence-transformers==2.2.2
google-cloud-aiplatform==1.40.0
streamlit==1.29.0
python-dotenv==1.0.0
numpy==1.24.3
pandas==2.1.4
scikit-learn==1.3.2
plotly==5.17.0
networkx==3.2.1
streamlit-chat==0.1.1
